+++ client/src/pages/splash/Splash.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {Link} from 'react-router-dom';
-import box from "../../../images/box5.png";
+import box from "../../images/box5.png";
 import "./splash.css";
 
 function Splash() {
